+++ database/migrations/2024_10_03_174453_create_stock_movements.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('stock_movements', function (Blueprint $table) {
             $table->integer('id')->primary();
-            $table->unsignedBigInteger('product_id');
+            $table->integer('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->unsignedBigInteger('quantity');
             $table->enum('type', ['input', 'output', 'return', 'loss']);
             $table->timestamp('created_at')->useCurrent();
-            $table->unsignedBigInteger('user_id');
+            $table->integer('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
